clean out code
