+++ bot/libs/utils/__init__.py
@@ -1,9 +1,6 @@
 from .embeds import (
-    ConfirmEmbed as ConfirmEmbed,
     Embed as Embed,
     ErrorEmbed as ErrorEmbed,
-    SuccessEmbed as SuccessEmbed,
-    TimeoutEmbed as TimeoutEmbed,
 )
 from .handler import KeyboardInterruptHandler as KeyboardInterruptHandler
 from .logger import CatherineLogger as CatherineLogger
@@ -11,5 +8,4 @@ from .modal import CatherineModal as CatherineModal
 from .reloader import Reloader as Reloader
 from .time import human_timedelta as human_timedelta
 from .tree import CatherineCommandTree as CatherineCommandTree
-from .utils import is_docker as is_docker
 from .view import CatherineView as CatherineView
