+++ bot/libs/utils/embeds.py
@@ -11,16 +11,6 @@ class Embed(discord.Embed):
         super().__init__(**kwargs)
 
 
-class SuccessEmbed(discord.Embed):
-    """Kumiko's custom success action embed"""
-
-    def __init__(self, **kwargs):
-        kwargs.setdefault("color", discord.Color.from_rgb(75, 181, 67))
-        kwargs.setdefault("title", "Action successful")
-        kwargs.setdefault("description", "The action requested was successful")
-        super().__init__(**kwargs)
-
-
 class ErrorEmbed(discord.Embed):
     """Kumiko's custom error embed"""
 
@@ -50,22 +40,3 @@ class FullErrorEmbed(ErrorEmbed):
         ```{error_traceback}```
         """
         return desc
-
-
-class TimeoutEmbed(discord.Embed):
-    """Timed out embed"""
-
-    def __init__(self, **kwargs):
-        kwargs.setdefault("color", discord.Color.from_rgb(214, 6, 6))
-        kwargs.setdefault("title", "\U00002757 Timed Out")
-        kwargs.setdefault("description", "Timed out waiting for a response. Cancelling action.")
-        super().__init__(**kwargs)
-
-
-class ConfirmEmbed(discord.Embed):
-    """Kumiko's custom confirm embed"""
-
-    def __init__(self, **kwargs):
-        kwargs.setdefault("color", discord.Color.from_rgb(255, 191, 0))
-        kwargs.setdefault("title", "Are you sure?")
-        super().__init__(**kwargs)
