+++ bot/libs/utils/pages/__init__.py
@@ -1,9 +1,7 @@
-from .base_pages import SimplePages as SimplePages
 from .paginator import (
     CatherineContextPages as CatherineContextPages,
     CatherinePages as CatherinePages,
 )
 from .sources import (
-    EmbedListSource as EmbedListSource,
     SimplePageSource as SimplePageSource,
 )
