--- bot/libs/utils/utils.py
@@ -1,8 +1,0 @@
-import os
-
-
-def is_docker() -> bool:
-    path = "/proc/self/cgroup"
-    return os.path.exists("/.dockerenv") or (
-        os.path.isfile(path) and any("docker" in line for line in open(path))
-    )
